move flusher around
